pyproject.toml
[tool.poetry]
name = "fastapi-sub-modules"
version = "0.1.0"
description = ""
authors = ["Abdullah Alqahtani <anqorithm@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
fastapi-sub-modules-logger-kit = { path = "./logger_kit", develop = true }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
